# Improve LVQ distance computation perf