--- a/src/vectors/lvq.rs
+++ b/src/vectors/lvq.rs
@@ -146,18 +146,12 @@ impl<'a, const B: usize> PrimaryVector<'a, B> {
     }
 
     fn dot_unnormalized(&self, other: &Self) -> f64 {
-        let dot_quantized = if B == 1 {
-            self.vector
-                .iter()
-                .zip(other.vector.iter())
-                .map(|(s, o)| (*s & *o).count_ones())
-                .sum::<u32>()
-        } else {
-            packing::unpack_iter::<B>(self.vector)
-                .zip(packing::unpack_iter::<B>(other.vector))
-                .map(|(s, o)| s as u32 * o as u32)
-                .sum::<u32>()
-        };
+        let dot_quantized = self
+            .vector
+            .iter()
+            .zip(other.vector.iter())
+            .map(|(s, o)| Self::dot_packed(*s, *o))
+            .sum::<u32>();
         let sdelta = f64::from(self.delta);
         let slower = f64::from(self.header.lower);
         let odelta = f64::from(other.delta);
@@ -166,6 +160,24 @@ impl<'a, const B: usize> PrimaryVector<'a, B> {
             + self.header.component_sum as f64 * sdelta * olower
             + other.header.component_sum as f64 * odelta * slower
             + slower * olower * (self.vector.len() * 8).div_ceil(B) as f64
+    }
+
+    fn dot_packed(a: u8, b: u8) -> u32 {
+        match B {
+            1 => (a & b).count_ones(),
+            2 => {
+                let a = [a & 0x3, (a >> 2) & 0x3, (a >> 4) & 0x3, (a >> 6) & 0x3];
+                let b = [b & 0x3, (b >> 2) & 0x3, (b >> 4) & 0x3, (b >> 6) & 0x3];
+                (a[0] * b[0] + a[1] * b[1] + a[2] * b[2] + a[3] * b[3]).into()
+            }
+            4 => {
+                let a = [a & 0xf, a >> 4];
+                let b = [b & 0xf, b >> 4];
+                (a[0] as u16 * b[0] as u16 + a[1] as u16 * b[1] as u16).into()
+            }
+            8 => a as u32 * b as u32,
+            _ => unimplemented!(),
+        }
     }
 }