… src/vectors/lvq/scalar.rs