Some of the changes are in scalar code and should be durable across platforms. Most of them are the result of
unpacking bit packed formats using SIMD intrinsics and shuffling.

Most modes improve by at least 65% with the exception of a previously optimized lvq1x1 path, some improve
by as much as 95%. Query (f32 x quantized) paths generally improve to the point where they are +/- 25% of
raw f32 dot product.

On an M2 Mac:
```
lvq2x8x8/doc/dot        time:   [331.65 ns 332.33 ns 333.16 ns]
                        change: [−95.942% −95.928% −95.914%] (p = 0.00 < 0.05)
lvq2x8x8/query/dot      time:   [289.43 ns 289.95 ns 290.51 ns]
                        change: [−68.261% −68.133% −67.973%] (p = 0.00 < 0.05)
lvq2x4x8/doc/dot        time:   [393.21 ns 395.79 ns 399.48 ns]
                        change: [−95.644% −95.586% −95.520%] (p = 0.00 < 0.05)
lvq2x4x8/query/dot      time:   [296.33 ns 297.15 ns 298.01 ns]
                        change: [−77.234% −76.682% −76.354%] (p = 0.00 < 0.05)
lvq2x4x4/doc/dot        time:   [459.80 ns 466.94 ns 477.01 ns]
                        change: [−94.773% −94.718% −94.638%] (p = 0.00 < 0.05)
lvq2x4x4/query/dot      time:   [299.99 ns 300.63 ns 301.35 ns]
                        change: [−77.165% −77.073% −76.977%] (p = 0.00 < 0.05)
lvq2x1x8/doc/dot        time:   [396.85 ns 398.56 ns 400.52 ns]
                        change: [−94.802% −94.777% −94.753%] (p = 0.00 < 0.05)
lvq2x1x8/query/dot      time:   [295.91 ns 296.79 ns 298.24 ns]
                        change: [−70.950% −70.838% −70.725%] (p = 0.00 < 0.05)
lvq1x8/doc/dot          time:   [27.340 ns 27.562 ns 27.970 ns]
                        change: [−26.905% −26.497% −25.905%] (p = 0.00 < 0.05)
lvq1x8/query/dot        time:   [249.37 ns 250.30 ns 251.19 ns]
                        change: [−73.041% −72.741% −72.489%] (p = 0.00 < 0.05)
lvq1x4/doc/dot          time:   [47.908 ns 48.075 ns 48.242 ns]
                        change: [−94.333% −94.311% −94.291%] (p = 0.00 < 0.05)
lvq1x4/query/dot        time:   [261.69 ns 262.36 ns 263.07 ns]
                        change: [−71.544% −71.435% −71.320%] (p = 0.00 < 0.05)
lvq1x1/doc/dot          time:   [10.236 ns 10.249 ns 10.263 ns]
                        change: [−2.9462% −2.5931% −2.2838%] (p = 0.00 < 0.05)
lvq1x1/query/dot        time:   [277.57 ns 282.43 ns 287.50 ns]
                        change: [−70.369% −69.993% −69.646%] (p = 0.00 < 0.05)
```